Refactored out authentication controls
