+++ web/src/features/authentication/index.tsx
@@ -1,0 +1,2 @@
+export * from './AuthenticationButton';
+export * from './Profile';
